move gen_env_var to new module
