+++ untitled/wsgi.py
@@ -11,6 +11,10 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "untitled.settings")
+from utils.env import get_env_var
+
+os.environ.setdefault(
+    "DJANGO_SETTINGS_MODULE",
+    "untitled.settings.{0}".format(get_env_var("RR_APP_MODE")))
 
 application = get_wsgi_application()
